+++ src/main/kotlin/dev/turingcomplete/intellijjvmsmanagerplugin/ui/detail/jvm/jvmaction/jtool/CreatingThreadDumpJvmAction.kt
@@ -1,10 +1,13 @@
 package dev.turingcomplete.intellijjvmsmanagerplugin.ui.detail.jvm.jvmaction.jtool
 
+import com.intellij.icons.AllIcons
 import com.intellij.openapi.Disposable
 import com.intellij.openapi.actionSystem.AnAction
 import com.intellij.openapi.project.Project
 import com.intellij.ui.components.JBCheckBox
-import com.intellij.unscramble.UnscrambleDialog
+import com.intellij.unscramble.AnalyzeStacktraceUtil
+import com.intellij.unscramble.ThreadDumpConsoleFactory
+import com.intellij.unscramble.ThreadDumpParser
 import com.intellij.util.ui.UIUtil
 import dev.turingcomplete.intellijjvmsmanagerplugin.ui.common.AnActionOptionButton
 import dev.turingcomplete.intellijjvmsmanagerplugin.ui.common.TextPopup
@@ -15,21 +18,19 @@ import dev.turingcomplete.intellijjvmsmanagerplugin.ui.detail.jvm.jvmaction.JvmA
 import java.awt.GridBagLayout
 import javax.swing.JPanel
 
+
 class CreatingThreadDumpJvmAction : JvmAction("Thread Dump") {
   // -- Companion Object -------------------------------------------------------------------------------------------- //
-
-  companion object {
-    private val TASK_TITLE : (JvmActionContext) -> String = {
-      "Creating Thread Dump of PID ${it.processNode.process.processID}"
-    }
-  }
-
   // -- Properties -------------------------------------------------------------------------------------------------- //
 
   private lateinit var printAdditionalLocksInformation: JBCheckBox
   private lateinit var  printAdditionalThreadsInformation: JBCheckBox
 
   private val createJToolCommand: (JvmActionContext) -> Pair<JTool, List<String>> = { createJToolCommand(it) }
+
+  private val toTaskTitleFunction : (JvmActionContext) -> String = {
+    "Creating Thread Dump of PID ${it.processNode.process.processID}"
+  }
 
   // -- Initialization ---------------------------------------------------------------------------------------------- //
   // -- Exposed Methods --------------------------------------------------------------------------------------------- //
@@ -53,12 +54,15 @@ class CreatingThreadDumpJvmAction : JvmAction("Thread Dump") {
   // -- Private Methods --------------------------------------------------------------------------------------------- //
 
   private fun createOpenResultInThreadDumpAnalyzerRunOption(): AnAction {
-    val onSuccess: (String, JvmActionContext) -> Unit = { result, jvmActionContext ->
-      val unscrambleDialog = UnscrambleDialog(jvmActionContext.project)
-      unscrambleDialog.setText(result)
-      unscrambleDialog.show()
+    val onSuccess: (String, JvmActionContext) -> Unit = { threadDump, jvmActionContext ->
+      val threadStates = ThreadDumpParser.parse(threadDump)
+      val project = jvmActionContext.project
+      val threadDumpConsoleFactory = if (threadStates.size > 1) ThreadDumpConsoleFactory(project, threadStates) else null
+      val tabTitle = "${jvmActionContext.processNode.smartName} (${jvmActionContext.processNode.process.processID})"
+      val icon = AllIcons.Actions.Dump
+      AnalyzeStacktraceUtil.addConsole(project, threadDumpConsoleFactory, tabTitle, threadDump, icon)
     }
-    return ToStringJToolRunOption("Run - Show In Thread Dump Analyzer", TASK_TITLE, createJToolCommand, onSuccess)
+    return ToStringJToolRunOption("Run - Show In Thread Dump Analyzer", toTaskTitleFunction, createJToolCommand, onSuccess)
   }
 
   private fun createOpenResultInPopupRunOption(): AnAction {
@@ -66,11 +70,11 @@ class CreatingThreadDumpJvmAction : JvmAction("Thread Dump") {
       val title = "Thread Dump of PID ${jvmActionContext.processNode.process.processID}"
       TextPopup.showCenteredInCurrentWindow(title, result, jvmActionContext.project, false)
     }
-    return ToStringJToolRunOption("Run - Show Output", TASK_TITLE, createJToolCommand, onSuccess)
+    return ToStringJToolRunOption("Run - Show Output", toTaskTitleFunction, createJToolCommand, onSuccess)
   }
 
   private fun createToFileJToolRunOption(): AnAction {
-    return ToFileJToolRunOption(TASK_TITLE, "thread-dump_", createJToolCommand)
+    return ToFileJToolRunOption(toTaskTitleFunction, "thread-dump_", createJToolCommand)
   }
 
   private fun createJToolCommand(jvmActionContext: JvmActionContext): Pair<JTool, List<String>> {